Почасовой прогноз готов

+++ app/src/main/java/com/example/mobilki/presentation/screens/weather/WeatherScreenSideEffect.kt
@@ -3,5 +3,5 @@ package com.example.mobilki.presentation.screens.weather
 sealed class WeatherScreenSideEffect {
     object RequestGpsPermission: WeatherScreenSideEffect()
 
-    class GoToHoursForecast(val lat:  Float, val lon: Float): WeatherScreenSideEffect()
+    class GoToHoursForecast(val lat:  Double, val lon: Double) : WeatherScreenSideEffect()
 }
